Bot base/low_latency_execution.py
import websocket
import threading
import asyncio
import aiohttp
from binance.client import Client
# --- INICIO DE MODIFICACIÓN: (Paso 3) ---
from unified_logger import get_logger
# --- FIN DE MODIFICACIÓN ---
import json 

# --- INICIO DE MODIFICACIÓN: (Paso 3) ---
logger = get_logger(__name__)
# --- FIN DE MODIFICACIÓN ---

# --- INICIO DE MODIFICACIÓN: Importar websockets y asyncio correctamente ---
# (El informe menciona 'websockets' y 'asyncio' como dependencias)
try:
    import websockets
except ImportError:
    logger.warning("Módulo 'websockets' no encontrado. La conexión directa fallará.", extra={'symbol': 'WS_IMPORT_ERROR'})
    websockets = None
# --- FIN DE MODIFICACIÓN ---


class LowLatencyExecution:
    def __init__(self):
        self.ws_connections = {}
        self.last_prices = {}
        self.execution_cache = {}
        
    async def direct_websocket_connection(self, symbols):
        """Conexión directa WebSocket a Binance para múltiples símbolos"""
        if websockets is None:
             logger.error("Módulo 'websockets' no instalado. No se puede iniciar direct_websocket_connection.", extra={'symbol': 'WS_EXEC'})
             return

        for symbol in symbols:
            ws_url = f"wss://stream.binance.com:9443/ws/{symbol.lower()}@depth@100ms"
            log_extra = {'symbol': symbol, 'transaction_id': 'WS_CONN'} # Log extra
            
            # --- INICIO DE MODIFICACIÓN: (Punto Crítico 2) ---
            try:
                # AÑADIR TIMEOUT
                async with websockets.connect(ws_url, ping_timeout=10, close_timeout=10) as websocket:
                    self.ws_connections[symbol] = websocket
                    
                    async for message in websocket:
                        data = json.loads(message)
                        await self.process_orderbook_update(symbol, data)
            except asyncio.TimeoutError:
                logger.error(f"Timeout en WebSocket para {symbol}", extra=log_extra)
            except Exception as e:
                logger.error(f"Error en WebSocket para {symbol}: {e}", exc_info=True, extra=log_extra)
            # --- FIN DE MODIFICACIÓN ---
    
    async def process_orderbook_update(self, symbol, data):
        """Procesa la actualización del orderbook (debe ser implementado)"""
        # Aquí es donde se actualizaría el orderbook local
        # logger.debug(f"Actualización de OB para {symbol}: {data}")
        self.execution_cache[symbol] = data # Guardar el último estado
        pass

    def get_current_orderbook(self, symbol):
        """Obtiene el orderbook más reciente del caché"""
        return self.execution_cache.get(symbol, {'bids': [], 'asks': []})

    def colocation_optimization(self):
        """Optimización de latencia de red y ejecución"""
        optimizations = {
            'tcp_nodelay': True,
            'websocket_ping_interval': 30,
            'websocket_ping_timeout': 10,
            'reconnect_delay': 1,
            'max_reconnect_attempts': 5,
            'use_compression': True
        }
        return optimizations
    
    def smart_order_routing(self, symbol, side, quantity, order_type="MARKET"):
        """
        Mejor ejecución usando múltiples order types y estrategias
        """
        log_extra = {'symbol': symbol, 'transaction_id': 'SOR'} # Log extra
        current_orderbook = self.get_current_orderbook(symbol)
        
        if not current_orderbook or not current_orderbook.get('bids') or not current_orderbook.get('asks'):
             logger.warning(f"No hay datos de orderbook en caché para {symbol}. Usando ejecución estándar.", extra=log_extra)
             # Aquí se debería llamar a la ejecución estándar (ej. bi.place_market_order)
             return None # Indicar que SOR falló

        if order_type == "MARKET":
            return self.execute_market_order(symbol, side, quantity, current_orderbook)
        elif order_type == "LIMIT_IOC":
            return self.execute_limit_ioc(symbol, side, quantity, current_orderbook)
        elif order_type == "TWAP":
            return self.execute_twap(symbol, side, quantity, duration=30)
        
        return None # Tipo de orden no soportado por SOR

    async def execute_market_order(self, symbol, side, quantity, orderbook):
        """Ejecución market order optimizada"""
        log_extra = {'symbol': symbol, 'transaction_id': 'SOR_MARKET'} # Log extra
        if side == "BUY":
            best_ask = float(orderbook['asks'][0][0])
            estimated_price = best_ask * 1.0005  # Pequeño slippage estimado
        else:
            best_bid = float(orderbook['bids'][0][0]) 
            estimated_price = best_bid * 0.9995
            
        logger.info(f"[SOR] Ejecutando MARKET {side} {quantity} {symbol} @ ~{estimated_price}", extra=log_extra)
        
        # Aquí iría la llamada real a la API (posiblemente aiohttp)
        
        return {
            'symbol': symbol,
            'side': side,
            'quantity': quantity,
            'estimated_price': estimated_price,
            'order_type': 'MARKET',
            'latency_optimized': True
        }

    async def execute_limit_ioc(self, symbol, side, quantity, orderbook):
        """Ejecución Limit IOC optimizada"""
        log_extra = {'symbol': symbol, 'transaction_id': 'SOR_IOC'} # Log extra
        logger.info(f"[SOR] Ejecutando LIMIT_IOC {side} {quantity} {symbol}", extra=log_extra)
        return {}

    async def execute_twap(self, symbol, side, quantity, duration):
        """Ejecución TWAP optimizada"""
        log_extra = {'symbol': symbol, 'transaction_id': 'SOR_TWAP'} # Log extra
        logger.info(f"[SOR] Ejecutando TWAP {side} {quantity} {symbol} por {duration}s", extra=log_extra)
        return {}